Moved local getters from FirebaseAccount to local DB-indipendent Account
